requirements.txt
Flask~=1.1.1
Flask-Cors==3.0.8
pymongo[tls,srv]
Werkzeug==0.16.0
pandas~=1.1.5
beautifulsoup4~=4.9.1
bs4~=0.0.1
requests~=2.27.1
sqlalchemy~=1.4.22
pytube~=12.1.0
boto3~=1.21.32
google-api-python-client~=2.60.0
snowflake-sqlalchemy~=1.4.1
gunicorn==20.1.0
markupsafe==2.0.1
dnspython==2.2.1